requirements/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.5.1/constraints-3.10.txt"

apache-airflow-providers-snowflake==4.0.2
apache-airflow-providers-mysql==4.0.0
apache-airflow-providers-apache-spark==4.0.0
pyarrow==9.0.0
numpy==1.22.4
pandas==1.5.2
pyspark==3.3.1
pyspark-pandas==0.0.7
openpyxl==3.1.2
